requirements.txt
Flask
Werkzeug
pdfplumber
python-dotenv
openai==1.2.3
tiktoken